+++ phpBB2/album_mod/album_hierarchy_sql.php
@@ -875,13 +875,14 @@ function album_get_moderator_info($cat) {
     // Modifications are done by IdleVoid
     $moderators = '';
     $grouprows = array();
+	$moderator_group_ids = album_sql_id_list($cat['cat_moderator_groups']);
 
     // We have usergroup_ID, now we need usergroup name
     $sql = "SELECT group_id, group_name
 			FROM " . GROUPS_TABLE . "
 			WHERE group_single_user <> 1
 				AND group_type <> " . GROUP_HIDDEN . "
-				AND group_id IN (" . $cat['cat_moderator_groups'] . ")
+				AND group_id IN ($moderator_group_ids)
 			ORDER BY group_name ASC";
 
 	if (!$result = $db->sql_query($sql)) 
@@ -900,7 +901,7 @@ function album_get_moderator_info($cat) {
 	{
 		for ($j = 0; $j < count($grouprows); $j++) 
 		{
-        	$group_link = '<a href="' . append_sid("groupcp.$phpEx?" . POST_GROUPS_URL . '=' . $grouprows[$j]['group_id']) . '">' . $grouprows[$j]['group_name'] . '</a>';
+			$group_link = '<a href="' . append_sid("groupcp.$phpEx?" . POST_GROUPS_URL . '=' . (int) $grouprows[$j]['group_id']) . '">' . album_html_text($grouprows[$j]['group_name']) . '</a>';
             $moderators .= ($moderators == '') ? $group_link : ', ' . $group_link;
         }
     }